feat(templates/base.html): add "About" link to the navigation bar style(templates/base.html): reformat script tags and improve code readability perf(templates/base.html): add integrity attribute to marked.js script for security
The "About" link has been added to the navigation bar to provide users with easy access to information about the site. The script tags have been reformatted to improve the readability of the code, making it easier to maintain. Additionally, an integrity attribute has been added to the marked.js script tag to ensure the security of the script being loaded, protecting against CDN tampering.